add menu system and a rom loader (#4)
* execute render in the main thread

* rm

* mess around with drawing on top of the nes display

* start a menu system

* menu system is interactable

* draw some snow in the menu

* have each snow fall at a different rate

* pass the window size to the snow generator

* dont react to emulator buttons while the menu is running. draw a simple button around the text

* move menu system to its own module

* toggle sound via the menu

* lift some functions to the top level

* try to separate input/output channels. clean up key handling

* ignore full channel

* pass the nesfile as a parameter

* start rom loader

* save every 60th frame

* move rom loader to the menu

* rom loader preview works

* render the thumbnails in rows

* sort the roms by their filename

* show thumbnails immediately

* increase generator size

* render snow behind menu. dont render thumbnails in blend mode

* move waiter.Add

* store the roms in sorted order on add

* avoid blocking the menu thread

* change the selected rom

* implement selecting a rom to load it

* show the current selected rom name

* move up/down in the menu. clean up renderUpdates usage

* tolerate not sending a window update

* log a warning

* use a separate goroutine for managing running the nes

* abort the rom loader early

* draw the name of the rom under the thumbnail

* compare using the base name only

* can scroll the tiles if there are too many

* use the rom id in the sort key so duplicate file names can be distinguished

* make sort key string weird

* move filename slightly

* draw an up/down arrow if there are more roms to show

* move font to data

* move font to data

* change reference to data

* cache textures

* cleanup

* cache the arrow texture

* use a common layout object for tiles. increase rom font size

* use constants for nes width/height/overscan

* show roms found so far

* use constant

+++ lib/mapper.go
@@ -232,7 +232,7 @@ func (mapper *Mapper1) Write(cpu *CPUState, address uint16, value byte) error {
             } else if address >= 0xc000 && address <= 0xdfff {
                 /* chr bank 1 */
                 if mapper.chrBankMode == 1 {
-                    base := uint16(mapper.register) * 0x1000
+                    base := uint32(mapper.register) * 0x1000
                     if len(mapper.CharacterMemory) != 0 {
                         if int(base + 0x1000) < len(mapper.CharacterMemory) {
                             cpu.PPU.CopyCharacterRom(0x1000, mapper.CharacterMemory[base:base + 0x1000])
